refactor: add interface
