+++ packages/axios/src/index.ts
@@ -1,2 +1,3 @@
 export { AutoConfiguration as Configuration } from './configuration';
+export * from './interface';
 export * from './servicManager';
