+++ packages/axios/src/servicManager.ts
@@ -1,55 +1,7 @@
-import axios, {
-  AxiosRequestConfig,
-  AxiosInstance,
-  AxiosResponse,
-  AxiosInterceptorManager,
-} from 'axios';
+import axios, { AxiosRequestConfig, AxiosInstance, AxiosResponse } from 'axios';
 import { Provide, Inject, Init, Config, App } from '@midwayjs/decorator';
 import { IMidwayApplication } from '@midwayjs/core';
-
-export interface AxiosHttpService {
-  interceptors: {
-    request: AxiosInterceptorManager<AxiosRequestConfig>;
-    response: AxiosInterceptorManager<AxiosResponse>;
-  };
-  getUri(config?: AxiosRequestConfig): string;
-  request<T = any, R = AxiosResponse<T>>(
-    config: AxiosRequestConfig
-  ): Promise<R>;
-  get<T = any, R = AxiosResponse<T>>(
-    url: string,
-    config?: AxiosRequestConfig
-  ): Promise<R>;
-  delete<T = any, R = AxiosResponse<T>>(
-    url: string,
-    config?: AxiosRequestConfig
-  ): Promise<R>;
-  head<T = any, R = AxiosResponse<T>>(
-    url: string,
-    config?: AxiosRequestConfig
-  ): Promise<R>;
-  options<T = any, R = AxiosResponse<T>>(
-    url: string,
-    config?: AxiosRequestConfig
-  ): Promise<R>;
-  post<T = any, R = AxiosResponse<T>>(
-    url: string,
-    data?: any,
-    config?: AxiosRequestConfig
-  ): Promise<R>;
-  put<T = any, R = AxiosResponse<T>>(
-    url: string,
-    data?: any,
-    config?: AxiosRequestConfig
-  ): Promise<R>;
-  patch<T = any, R = AxiosResponse<T>>(
-    url: string,
-    data?: any,
-    config?: AxiosRequestConfig
-  ): Promise<R>;
-}
-
-const AXIOS_INSTANCE_KEY = 'axios:instance';
+import { AXIOS_INSTANCE_KEY, AxiosHttpService } from './interface';
 
 @Provide()
 export class HttpService implements AxiosHttpService {
@@ -114,35 +66,41 @@ export class HttpService implements AxiosHttpService {
   getUri(config?: AxiosRequestConfig): string {
     return this.instance.getUri(config);
   }
+
   request<T = any, R = AxiosResponse<T>>(
     config: AxiosRequestConfig
   ): Promise<R> {
     return this.instance.request(config);
   }
+
   get<T = any, R = AxiosResponse<T>>(
     url: string,
     config?: AxiosRequestConfig
   ): Promise<R> {
     return this.instance.get(url, config);
   }
+
   delete<T = any, R = AxiosResponse<T>>(
     url: string,
     config?: AxiosRequestConfig
   ): Promise<R> {
     return this.instance.delete(url, config);
   }
+
   head<T = any, R = AxiosResponse<T>>(
     url: string,
     config?: AxiosRequestConfig
   ): Promise<R> {
     return this.instance.head(url, config);
   }
+
   options<T = any, R = AxiosResponse<T>>(
     url: string,
     config?: AxiosRequestConfig
   ): Promise<R> {
     return this.instance.options(url, config);
   }
+
   post<T = any, R = AxiosResponse<T>>(
     url: string,
     data?: any,
@@ -150,6 +108,7 @@ export class HttpService implements AxiosHttpService {
   ): Promise<R> {
     return this.instance.post(url, config);
   }
+
   put<T = any, R = AxiosResponse<T>>(
     url: string,
     data?: any,
@@ -157,6 +116,7 @@ export class HttpService implements AxiosHttpService {
   ): Promise<R> {
     return this.instance.put(url, config);
   }
+
   patch<T = any, R = AxiosResponse<T>>(
     url: string,
     data?: any,
